Refactor + test
