+++ tests/test_cli.py
@@ -4,7 +4,16 @@ from unittest.mock import Mock, patch
 from click.testing import CliRunner
 from requests import RequestException
 
+from cert_host_scraper import __version__
 from cert_host_scraper.cli import cli
+
+
+class TestVersion(TestCase):
+    def test_version(self):
+        runner = CliRunner()
+        result = runner.invoke(cli, ["--version"])
+        self.assertEqual(result.exit_code, 0)
+        self.assertEqual(result.output, f"{__version__}\n")
 
 
 class TestSearch(TestCase):
@@ -13,7 +22,7 @@ class TestSearch(TestCase):
         result = runner.invoke(cli, ["search"])
         self.assertEqual(result.exit_code, 2)
 
-    @patch("cert_host_scraper.fetch_urls")
+    @patch("cert_host_scraper.cli.fetch_urls")
     def test_search_network_error(self, mock_fetch_urls: Mock):
         mock_fetch_urls.side_effect = RequestException()
         runner = CliRunner()
